requirements.txt
streamlit
langchain
tabulate
faiss-cpu
python-dotenv